change i32 with i64 in ProtocolType::integer

+++ src/protocol/parser/integer.rs
@@ -5,7 +5,7 @@ use crate::protocol::types::ProtocolType;
 /// Parses a serialized RESP signed integer into a ProtocolType::Integer
 ///
 pub struct IntegerParser {
-    data: i32,
+    data: i64,
 }
 
 impl IntegerParser {
